+++ frontend/src/components/ui/icon/CrossIcon.tsx
